web/requirements.txt
# alembic==1.6.5
anytree==2.8.0
# asgiref==3.6.0
# async-generator==1.10
# async-timeout==4.0.2
# attrs==21.4.0
# Babel==2.9.1
# beautifulsoup4==4.11.2
# blinker==1.4
# certifi==2021.5.30
# chardet==4.0.0
# charset-normalizer==2.0.12
# click==8.0.1
# commonmark==0.9.1
crispy-bootstrap4==2022.1
# cryptography==37.0.2
# cycler==0.11.0
# defusedxml==0.7.1
# Deprecated==1.2.13
Django==4.1
django-appconf==1.0.5
django-ckeditor==6.5.1
django-cleanup==7.0.0
django-crispy-forms==2.0
django-cron==0.6.0
django-crontab==0.7.1
django-extensions==3.2.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-phone-field==1.8.1
django-redis==5.2.0
django-select2==8.1.1
dnspython==2.1.0
dominate==2.6.0
elasticsearch==7.13.3
email-validator==1.1.3
et-xmlfile==1.1.0
graphkit==1.2.4
graphviz==0.20.1
greenlet==1.1.0
guess-language-spirit==0.5.3
gunicorn==20.1.0
h11==0.13.0
httpie==3.2.1
httplib2==0.11.3
idna==2.10
# importlib-resources==5.7.1
isodate==0.6.1
# itsdangerous==2.0.1
# Jinja2==3.0.1
jsonmerge==1.7.0
# jwt==1.3.1
kiwisolver==1.4.2
langdetect==1.0.9
lxml==4.8.0
# Mako==1.1.4
mapper==0.1.17
# MarkupSafe==2.0.1
matplotlib==3.3.3
multidict==6.0.2
# networkx==2.6.3
numpy==1.22.3
openpyxl==3.0.7
outcome==1.1.0
packaging==21.3
pandas==1.4.2
pathlib==1.0.1
pdf2image==1.16.0
Pillow==8.0.1
ply==3.11
psycopg2-binary==2.8.6
pycel==1.0b30
pydot==1.4.2
pydotplus==2.0.2
# Pygments==2.9.0
# PyJWT==2.1.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2021.1
rdflib==6.1.1
redis==4.3.1
requests==2.25.1
requests-toolbelt==0.9.1
rich==12.4.1
rq==1.9.0
# ruamel.yaml==0.17.21
# ruamel.yaml.clib==0.2.7
scipy==1.5.4
# selenium==4.2.0
six==1.16.0
# sniffio==1.2.0
# sortedcontainers==2.4.0
# soupsieve==2.4
# SPARQLWrapper==2.0.0
SQLAlchemy==1.4.20
sqlparse==0.4.3
tqdm==4.54.0
trio==0.21.0
trio-websocket==0.9.2
Unidecode==1.2.0
urllib3==1.26.9
visitor==0.1.3
webcolors==1.13
webdriver-manager==3.7.0
webscrapbook==1.6.0
Werkzeug==2.0.1
wrapt==1.14.1
wsproto==1.1.0
WTForms==2.3.3
xlrd==2.0.1
zipp==3.8.0
Faker
sphinx
sphinxcontrib-httpdomain